adding features for like and dislike now edit

+++ Backend/root/blog/api/views.py
@@ -26,19 +26,23 @@ class BlogDetail_GV(RetrieveUpdateDestroyAPIView):
     queryset = Blog.objects.all()
     serializer_class = BlogSerializer
 
-@api_view(['POST'])
+@api_view(['GET', 'POST'])
 @authentication_classes([TokenAuthentication])
 @permission_classes([IsAuthenticated])
 def toggle_like(request, post_id):
     blog = get_object_or_404(Blog, id=post_id)
     user = request.user
-
-    try:
-        like = Like.objects.get(user=user, blog=blog)
-        like.delete()
-        liked = False
-    except Like.DoesNotExist:
-        Like.objects.create(user=user, blog=blog)
-        liked = True
-    data = {'liked': liked, } # 'likes_count': blog.like_set.count()
-    return JsonResponse(data)
+    if request.method == "GET":
+        liked = Like.objects.filter(user=user, blog=blog).exists()
+        data = {'liked': liked, }
+        return JsonResponse(data)
+    if request.method == "POST":
+        try:
+            like = Like.objects.get(user=user, blog=blog)
+            like.delete()
+            liked = False
+        except Like.DoesNotExist:
+            Like.objects.create(user=user, blog=blog)
+            liked = True
+        data = {'liked': liked, } # 'likes_count': blog.like_set.count()
+        return JsonResponse(data)
